ashtech: explicitly cast NAN to double
Compile error on MacOS:
In file included from ../../src/drivers/gps/devices/src/ashtech.cpp:41:

../../src/drivers/gps/devices/src/ashtech.h:114:75: fatal error: implicit conversion increases floating-point precision: 'float' to 'double' [-Wdouble-promotion]

        void sendSurveyInStatusUpdate(bool active, bool valid, double latitude = NAN, double longitude = NAN,

                                                                               ~ ^~~

/usr/include/math.h:60:28: note: expanded from macro 'NAN'

#   define    NAN          __builtin_nanf("0x7fc00000")

+++ src/ashtech.cpp
@@ -1119,7 +1119,7 @@ int GPSDriverAshtech::configure(unsigned &baudrate, OutputMode output_mode)
 
 	if (output_mode == OutputMode::RTCM && _board == AshtechBoard::trimble_mb_two) {
 		SurveyInStatus status;
-		status.latitude = status.longitude = NAN;
+		status.latitude = status.longitude = (double)NAN;
 		status.altitude = NAN;
 		status.duration = 0;
 		status.mean_accuracy = 0;
